backend/project/serializers.py
from rest_framework import serializers

from accounts.models import User
from .models import Project

class ProjectSerializer(serializers.ModelSerializer):
    class Meta:
        model = Project
        fields = "__all__"
        read_only_fields = ("slug",)