backend/requirements.txt
Django==4.2.2
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
django-cors-headers==4.1.0
gunicorn==20.1.0
whitenoise==6.4.0
django-environ==0.10.0
django-filter==23.2